Create Capybara integration specs with FactoryGirl for user create, edit, and delete. Update models and schema to change number from integer to bigint.

+++ db/migrate/20150401172526_devise_create_users.rb
@@ -2,7 +2,7 @@ class DeviseCreateUsers < ActiveRecord::Migration
   def change
     create_table(:users) do |t|
       t.string :name
-      t.integer :number
+      t.bigint :number
 
       ## Database authenticatable
       t.string :email,              null: false, default: ""
